feat(collection_proxy): introduce proxy for the native Collection
Instead of creating a new `Collection` (and `Cursor`) class in our
module, we can instead simply wrap the native driver's `Collection`
and trap method calls to it. The reason we had to wrap it in the
first place is because custom render methods were injected, but
ES6's proxy let's us get away with this without all the boilerplate
code involved.