refactor(repl): drop `preprocess` in favor of using translator
For the REPL translation case we used to use a regular expression
for dynamic async IIFE wrapping for assignment. Instead, this has
been moved over to our active translation unit. Since we already
incur the overhead of parsing/walking the AST, it makes sense to
do it here, since we have far more context to work with.