refactor(repl): attempt initial parsing not using acorn
There was a bug in our previous preparse for Recoverable errors
when there were comments involved in the input. Because we didn't
attempt parse in a loop, it ended prematurely and fed acorn
incomplete data. Instead, we now do what the node repl does and
use the vm module to parse before attempting a rewrite.